Move numeric tests into the numeric test submodule
